fix: TODO V4 — Decimal financier, merge lock, email sanitize, soft-delete, frontend bugs
CRITIQUE:
- Modeles financiers: Mapped[float] → Mapped[Decimal] (payment, devis, facture, pec)
- Race condition merge: acquire_lock distribue avant fusion clients
- Email header injection: subject sanitise (supprime \r\n\t)

ELEVE:
- client_repo.get_by_id() filtre deleted_at par defaut (12 services proteges)
- get_by_id_including_deleted() pour restore/admin
- InlineEdit: setTimeout cleanup sur unmount
- Token refresh: clearTimeout avant retry
- SSE: flag isReconnecting anti-boucle
- Popstate: history.go(1) au lieu de pushState (corrige back button trap)

MOYEN:
- SWR dedupingInterval: 5000 → 2000ms
- Keyboard shortcuts: commentaire intentionnel pour full reload
- useTransition: reset showDuplicates dans catch
- cosium_document_sync: batch commit/50 au lieu de per-document
- CELERY_WORKER + SEED_ON_STARTUP migres dans Settings Pydantic
- .env.example mis a jour

+++ backend/app/models/ai.py
@@ -1,4 +1,5 @@
 from datetime import UTC, datetime
+from decimal import Decimal
 
 from sqlalchemy import DateTime, ForeignKey, Integer, Numeric, String
 from sqlalchemy.orm import Mapped, mapped_column
@@ -15,7 +16,7 @@ class AiUsageLog(Base):
     model_used: Mapped[str] = mapped_column(String(100), nullable=False)
     tokens_in: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
     tokens_out: Mapped[int] = mapped_column(Integer, nullable=False, default=0)
-    cost_usd: Mapped[float] = mapped_column(Numeric(10, 6), nullable=False, default=0)
+    cost_usd: Mapped[Decimal] = mapped_column(Numeric(10, 6), nullable=False, default=0)
     created_at: Mapped[datetime] = mapped_column(
         DateTime, default=lambda: datetime.now(UTC), nullable=False, index=True
     )
